bot/requirements.txt
aiohttp>=3.7.4
discord.py>=1.7.2
python-dotenv
